Add loader for texture
